backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

redis[hiredis]==5.0.1

python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
cryptography==42.0.2

httpx==0.26.0
yagmail==0.15.293
anthropic>=0.40.0
twilio>=8.0.0

pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1
pytz==2024.1

websockets==12.0
